pip-requirements.txt
# This file lists CKAN's dependencies so that you can install them (e.g. into
# a Python virtual environment) after you have installed CKAN, with a command
# like: pip install -r pip-requirements.txt. See the Install from Source
# instructions in CKAN's documentation for full installation instructions.
Genshi==0.6
sqlalchemy-migrate==0.7.2
sqlalchemy==0.7.8
webhelpers==1.3
pyutilib.component.core==4.5.3
-e git+https://github.com/okfn/vdm.git@vdm-0.11#egg=vdm
solrpy==0.9.5
formalchemy==1.4.2
pairtree==0.7.1-T
ofs==0.4.1
apachemiddleware==0.1.1
markupsafe==0.15
babel==0.9.6
psycopg2==2.4.5
webob==1.0.8
Pylons==0.9.7
repoze.who==1.0.19
tempita==0.5.1
zope.interface==4.0.1
repoze.who.plugins.openid==0.5.3
repoze.who-friendlyform==1.0.8
routes==1.13
paste==1.7.5.1
Jinja2==2.6
fanstatic==0.12
